feat: 게시글 좋아요 기능

+++ community/src/main/java/com/example/community/dto/posts/PagingMetaResponse.java
@@ -1,14 +1,15 @@
 package com.example.community.dto.posts;
 
 public class PagingMetaResponse {
-    private int pageNumber;
-    private int pageSize = 20;
-    private int totalPosts;
-    private int totalPages;
+    private Long pageNumber;
+    private Long pageSize = 20L;
+    private Long totalPosts;
+    private Long totalPages;
     private String sortCondition = "createdAt,desc";
 
     public PagingMetaResponse(){}
-    public PagingMetaResponse(int pageNumber, int pageSize, int totalPosts, int totalPages, String sortCondition) {
+    public PagingMetaResponse(Long pageNumber, Long pageSize, Long totalPosts,
+                              Long totalPages, String sortCondition) {
         this.pageNumber = pageNumber;
         this.pageSize = pageSize;
         this.totalPosts = totalPosts;
@@ -16,14 +17,14 @@ public class PagingMetaResponse {
         this.sortCondition = sortCondition;
     }
 
-    public int getPageNumber() {return pageNumber;}
-    public void setPageNumber(int pageNumber) {this.pageNumber = pageNumber;}
-    public int getPageSize() {return pageSize;}
-    public void setPageSize(int pageSize) {this.pageSize = pageSize;}
-    public int getTotalPosts() {return totalPosts;}
-    public void setTotalPosts(int totalPosts) {this.totalPosts = totalPosts;}
-    public int getTotalPages() {return totalPages;}
-    public void setTotalPages(int totalPages) {this.totalPages = totalPages;}
+    public Long getPageNumber() {return pageNumber;}
+    public void setPageNumber(Long pageNumber) {this.pageNumber = pageNumber;}
+    public Long getPageSize() {return pageSize;}
+    public void setPageSize(Long pageSize) {this.pageSize = pageSize;}
+    public Long getTotalPosts() {return totalPosts;}
+    public void setTotalPosts(Long totalPosts) {this.totalPosts = totalPosts;}
+    public Long getTotalPages() {return totalPages;}
+    public void setTotalPages(Long totalPages) {this.totalPages = totalPages;}
     public String getSortCondition() {return sortCondition;}
     public void setSortCondition(String sortCondition) {this.sortCondition = sortCondition;}
 }
